+++ source/javascripts/pages/TriggersPage/components/LegacyTriggers/AddPrTriggerDialog.tsx
@@ -3,12 +3,13 @@ import {
   Box,
   Button,
   Card,
-  DefinitionTooltip,
+  Checkbox,
   Dialog,
   DialogBody,
   DialogFooter,
   Divider,
   Input,
+  Link,
   ProgressIndicator,
   ProgressIndicatorProps,
   Select,
@@ -17,9 +18,9 @@ import {
 } from '@bitrise/bitkit';
 import { Controller, FormProvider, useFieldArray, useForm, useFormContext } from 'react-hook-form';
 
-import { Condition, FormItems, LegacyPushConditionType, TriggerItem } from './TriggersPage.types';
-import { checkIsConditionsUsed } from './TriggersPage.utils';
-import RegexCheckbox from './RegexCheckbox';
+import { Condition, FormItems, LegacyPrConditionType, TriggerItem } from '../TriggersPage/TriggersPage.types';
+import RegexCheckbox from '../SelectiveTriggers/RegexCheckbox';
+import { checkIsConditionsUsed } from '../TriggersPage/TriggersPage.utils';
 
 type DialogProps = {
   currentTriggers: TriggerItem[];
@@ -31,15 +32,18 @@ type DialogProps = {
   workflows: string[];
 };
 
-const LABEL_MAP: Record<LegacyPushConditionType, string> = {
-  push_branch: 'Push branch',
-  commit_message: 'Commit message',
-  changed_files: 'Path',
+const LABEL_MAP: Record<LegacyPrConditionType, string> = {
+  pull_request_target_branch: 'Enter a target branch',
+  pull_request_source_branch: 'Enter a source branch',
+  pull_request_label: 'Enter a label',
+  pull_request_comment: 'Enter a comment',
+  commit_message: 'Enter a commit message',
+  changed_files: 'Enter a path',
 };
 
-const getLabelText = (isRegex: boolean, type: LegacyPushConditionType): string => {
+const getLabelText = (isRegex: boolean, type: LegacyPrConditionType): string => {
   if (isRegex) {
-    return 'Regex pattern';
+    return 'Enter a regex pattern';
   }
   return LABEL_MAP[type];
 };
@@ -49,13 +53,16 @@ type ConditionCardProps = {
   conditionNumber: number;
 };
 
-const OPTIONS_MAP: Record<LegacyPushConditionType, string> = {
-  push_branch: 'Push branch',
+const OPTIONS_MAP: Record<LegacyPrConditionType, string> = {
+  pull_request_target_branch: 'Target branch',
+  pull_request_source_branch: 'Source branch',
+  pull_request_label: 'PR label',
+  pull_request_comment: 'PR comment',
   commit_message: 'Commit message',
   changed_files: 'File change',
 };
 
-const ConditionCard = (props: ConditionCardProps) => {
+export const ConditionCard = (props: ConditionCardProps) => {
   const { children, conditionNumber } = props;
   const { control, watch, setValue } = useFormContext();
   const { conditions } = watch();
@@ -108,9 +115,14 @@ const ConditionCard = (props: ConditionCardProps) => {
               />
             )}
           />
-          {type === 'push_branch' && (
+          {type === 'pull_request_target_branch' && (
             <Text color="sys/neutral/base" textStyle="body/sm/regular">
-              If you leave it blank, Bitrise will start builds for any push branch.
+              If you leave it blank, Bitrise will start builds for any target branch.
+            </Text>
+          )}
+          {type === 'pull_request_source_branch' && (
+            <Text color="sys/neutral/base" textStyle="body/sm/regular">
+              If you leave it blank, Bitrise will start builds for any source branch.
             </Text>
           )}
         </>
@@ -119,7 +131,7 @@ const ConditionCard = (props: ConditionCardProps) => {
   );
 };
 
-const AddPushTriggerDialog = (props: DialogProps) => {
+const AddPrTriggerDialog = (props: DialogProps) => {
   const { currentTriggers, isOpen, onClose, pipelines, onSubmit, editedItem, workflows } = props;
   const [activeStageIndex, setActiveStageIndex] = useState<0 | 1>(0);
 
@@ -138,13 +150,14 @@ const AddPushTriggerDialog = (props: DialogProps) => {
       conditions: [
         {
           isRegex: false,
-          type: 'push_branch',
+          type: 'pull_request_target_branch',
           value: '',
         },
       ],
       id: crypto.randomUUID(),
       pipelineable: '',
-      source: 'push',
+      source: 'pull_request',
+      isDraftPr: true,
       isActive: true,
       ...editedItem,
     };
@@ -155,7 +168,7 @@ const AddPushTriggerDialog = (props: DialogProps) => {
     defaultValues,
   });
 
-  const { control, reset, handleSubmit, watch } = formMethods;
+  const { control, formState, reset, handleSubmit, watch, setValue } = formMethods;
 
   useEffect(() => {
     reset(defaultValues);
@@ -197,13 +210,18 @@ const AddPushTriggerDialog = (props: DialogProps) => {
     });
   };
 
-  const { conditions, pipelineable } = watch();
+  const { conditions, pipelineable, isDraftPr } = watch();
 
-  const isConditionsUsed = checkIsConditionsUsed(currentTriggers, watch() as TriggerItem);
+  let isConditionsUsed = checkIsConditionsUsed(currentTriggers, watch() as TriggerItem);
+
+  // Because draft PR checkbox is a condition
+  if (formState.dirtyFields.isDraftPr) {
+    isConditionsUsed = false;
+  }
 
   let hasEmptyCondition = false;
   conditions.forEach(({ type, value }) => {
-    if ((type !== 'push_branch' && !value) || !type) {
+    if ((!(type === 'pull_request_target_branch' || type === 'pull_request_source_branch') && !value) || !type) {
       hasEmptyCondition = true;
     }
   });
@@ -216,7 +234,7 @@ const AddPushTriggerDialog = (props: DialogProps) => {
         as="form"
         isOpen={isOpen}
         onClose={onFormCancel}
-        title={isEditMode ? 'Edit trigger' : 'Add push trigger'}
+        title={isEditMode ? 'Edit trigger' : 'Add pull request trigger'}
         maxWidth="480"
         onSubmit={handleSubmit(onFormSubmit)}
       >
@@ -232,9 +250,9 @@ const AddPushTriggerDialog = (props: DialogProps) => {
               </Text>
               <Text color="text/secondary" marginBottom="24">
                 Configure the{' '}
-                <DefinitionTooltip label="Configure the conditions that should all be met to execute the targeted Pipeline or Workflow.">
+                <Tooltip label="Configure the conditions that should all be met to execute the targeted Pipeline or Workflow.">
                   conditions
-                </DefinitionTooltip>{' '}
+                </Tooltip>{' '}
                 that should all be met to execute the targeted Pipeline or Workflow.
               </Text>
               {fields.map((item, index) => {
@@ -253,11 +271,30 @@ const AddPushTriggerDialog = (props: DialogProps) => {
                 variant="secondary"
                 leftIconName="PlusAdd"
                 width="100%"
+                marginBottom="24"
                 onClick={onAppend}
                 isDisabled={fields.length >= Object.keys(OPTIONS_MAP).length}
               >
                 Add condition
               </Button>
+              <Checkbox
+                isChecked={isDraftPr}
+                helperText={
+                  <>
+                    Supported for GitHub and GitLab.{' '}
+                    <Link
+                      colorScheme="purple"
+                      href="https://devcenter.bitrise.io/en/builds/starting-builds/triggering-builds-automatically.html#triggering-builds-from-draft-prs"
+                      isExternal
+                    >
+                      Learn more
+                    </Link>
+                  </>
+                }
+                onChange={(e) => setValue(`isDraftPr`, e.target.checked)}
+              >
+                Include draft pull requests
+              </Checkbox>
             </>
           ) : (
             <>
@@ -335,4 +372,4 @@ const AddPushTriggerDialog = (props: DialogProps) => {
   );
 };
 
-export default AddPushTriggerDialog;
+export default AddPrTriggerDialog;
