+++ source/javascripts/pages/TriggersPage/components/LegacyTriggers/AddTagTriggerDialog.tsx
@@ -1,10 +1,9 @@
 import { useEffect, useMemo } from 'react';
 import { Button, Dialog, DialogBody, DialogFooter, Input, Select, Text, Tooltip } from '@bitrise/bitkit';
 import { Controller, FormProvider, useForm } from 'react-hook-form';
-import { checkIsConditionsUsed } from './TriggersPage.utils';
-
-import { FormItems, TriggerItem } from './TriggersPage.types';
-import RegexCheckbox from './RegexCheckbox';
+import { FormItems, TriggerItem } from '../TriggersPage/TriggersPage.types';
+import RegexCheckbox from '../SelectiveTriggers/RegexCheckbox';
+import { checkIsConditionsUsed } from '../TriggersPage/TriggersPage.utils';
 
 type DialogProps = {
   currentTriggers: TriggerItem[];
