+++ source/javascripts/pages/TriggersPage/components/SelectiveTriggers/RegexCheckbox.tsx
