+++ source/javascripts/pages/TriggersPage/components/TriggersPage/TriggersPage.tsx
@@ -5,9 +5,9 @@ import WindowUtils from '@/core/utils/WindowUtils';
 import RuntimeUtils from '@/core/utils/RuntimeUtils';
 import { BitriseYml } from '@/core/models/BitriseYml';
 import BitriseYmlProvider from '@/contexts/BitriseYmlProvider';
+import LegacyTriggers from '../LegacyTriggers/LegacyTriggers';
+import SelectiveTriggers from '../SelectiveTriggers/SelectiveTriggers';
 import { getPipelineableTriggers } from './TriggersPage.utils';
-import SelectiveTriggers from './components/SelectiveTriggers';
-import LegacyTriggers from './components/LegacyTriggers';
 
 const TRIGGERS_CONFIGURED_METADATA_KEY = 'wfe_triggers_configure_webhooks_notification_closed';
