+++ source/javascripts/pages/index.ts
@@ -1,6 +1,6 @@
 import PipelinesPage from './PipelinesPage/PipelinesPage';
 import SecretsPage from './SecretsPage/SecretsPage';
-import TriggersPage from './TriggersPage/TriggersPage';
+import TriggersPage from './TriggersPage/components/TriggersPage/TriggersPage';
 import WorkflowsPage from './WorkflowsPage/WorkflowsPage';
 
 export { PipelinesPage, SecretsPage, TriggersPage, WorkflowsPage };
